# Optimize performance of `character_length` function