@comphead 2024-12-10
Thanks everyone.  One optimized a lot more to go!