@comphead 2024-12-10
Slightly late: but probably `take_iter_unchecked` will give some gain instead of traversing arrays