@waynexia 2025-07-23
> Slightly late: but probably `take_iter_unchecked` will give some gain instead of traversing arrays

This method accepts `StringArrayType` instead of `StringArray` or similar things so we can only use `.value()` or `.iter()` here if we don't do upcast/downcast 😢